+++ react/QuoteCreate.tsx
@@ -106,6 +106,16 @@ const QuoteCreate: StorefrontFunctionComponent<WrappedComponentProps & any> = ({
         title: translateMessage({
           id: 'store/orderquote.cartList.label.itemName',
         }),
+        // eslint-disable-next-line react/display-name
+        cellRenderer: ({ rowData }: any) => {
+          return rowData.skuName !== rowData.name ? (
+            <div>
+              <span>{rowData.name}</span>
+            </div>
+          ) : (
+            rowData.skuName
+          )
+        },
         minWidth: 300,
       },
       sellingPrice: {
@@ -121,6 +131,7 @@ const QuoteCreate: StorefrontFunctionComponent<WrappedComponentProps & any> = ({
             </span>
           )
         },
+        width: 200,
       },
       quantity: {
         title: translateMessage({
@@ -143,6 +154,7 @@ const QuoteCreate: StorefrontFunctionComponent<WrappedComponentProps & any> = ({
             </span>
           )
         },
+        width: 300,
       },
     },
   }
@@ -355,7 +367,6 @@ const QuoteCreate: StorefrontFunctionComponent<WrappedComponentProps & any> = ({
           })
       }
       activeLoading(false)
-      toastMessage('store/orderquote.create.error')
     }
   }
 
@@ -462,7 +473,7 @@ const QuoteCreate: StorefrontFunctionComponent<WrappedComponentProps & any> = ({
                 fullWidth
                 schema={defaultSchema}
                 items={itemsCopy}
-                density="high"
+                density="medium"
               />
             </div>
           </div>